wrong travel time when adding in bins

+++ Nouna_PCVRP/qt-project/ver00/nouna_project/hrd/interviewer.h
@@ -8,7 +8,7 @@ struct stRoute
 {
     std::vector<unsigned int> households;
     std::vector<unsigned int> villages;
-    unsigned int time;
+    float time;
 };
 
 /*
